pyproject.toml
[project]
name = "eindjeboss"
version = "0.1.0"
description = "Utility Discord bot written for the Eindhoven Community Discord server"
authors = [
    {name = "Jamil"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "aiocron (>=2.1,<3.0)",
    "asyncpraw (>=7.8.1,<8.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "bing-image-downloader (>=1.1.2,<2.0.0)",
    "cinemagoer (>=2023.5.1,<2024.0.0)",
    "colorthief (>=0.2.1,<0.3.0)",
    "dateparser (>=1.2.1,<2.0.0)",
    "deepl (>=1.22.0,<2.0.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "exceptiongroup (>=1.3.0,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "spotipy (>=2.25.1,<3.0.0)",
    "table2ascii (>=1.1.3,<2.0.0)",
    "wikipediasummary (>=1.1.4,<2.0.0)",
    "peewee (>=3.18.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
